# 2damvi-server-1/src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "",
    "version": "1.0.0",
    "title": "NodePlayers"
  },
  "paths": {
    "/": {
      "get": {
        "tags": ["Show Players"],
        "summary": "Principal Screen",
        "description": "This is the principal screen",
        "parameters": [],
        "responses": {
          "100":{
            "schema":{
              "$ref":"#/definitions/codes"
            }
          }
        }
      }
    },
    "/players/{alias}": {
      "parameters": [
        {
          "name": "alias",
          "in": "path",
          "required": true,
          "description": "Alias of player that we want to find",
          "type": "string"
        }
      ],
      "get": {
        "tags": ["Show Players"],
        "summary": "Show Players",
        "description": "Show a specific player",
        "parameters": [
          {
            "name": "alias",
            "in": "path",
            "description": "The Alias of the Player.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "100":{
            "schema":{
              "$ref":"#/definitions/codes"
            }
          }
        }
      },
      "post": {
        "tags": ["Edit Players"],
        "summary": "Create Players",
        "description": "Enter a new player",
        "parameters": [
          {
            "name": "Player",
            "in": "body",
            "description": "Player that we want to create",
            "schema": {
              "$ref": "#/definitions/Player"
            }
          }
        ],
        "responses": {
          "100":{
            "schema":{
              "$ref":"#/definitions/codes"
            }
          }
        }
      },
      "put": {
        "tags": ["Edit Players"],
        "summary": "Edit a Player",
        "description": "Edit a player",
        "parameters": [
          {
            "name": "Player",
            "in": "body",
            "description": "Player that we want to update",
            "schema": {
              "$ref": "#/definitions/Player"
            }
          }
        ],
        "responses": {
          "100":{
            "schema":{
              "$ref":"#/definitions/codes"
            }
          }
        }
      },
      "delete":{
        "tags": ["Delete a Player"],
        "summary": "Delete a Player",
        "description": "Delete a specific player",
        "parameters": [
          {
            "name": "alias",
            "in": "path",
            "description": "The Alias of the Player.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "100":{
            "schema":{
              "$ref":"#/definitions/codes"
            }
          }
        }
      }
    },
    "/ranking": {
      "get": {
        "tags": ["Show Players"],
        "summary": "Show Players",
        "description": "Show all the players",
        "parameters": [],
        "responses": {
          "400": {
            "description": "OK",
            "schema": {
              "type": "object",
              "$ref":"#/definitions/Player"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Player":{
      "required": [
        "name",
        "surname",
        "score"
      ],
      "properties": {
        "name": {
          "type": "string",
          "uniqueItems": true
        },
        "surname": {
          "type": "string"
        },
        "score": {
          "type": "integer"
        }
      },
      "example":{
        "name": "Samuel",
        "surname": "Viguera",
        "score": 3
      }
    },
    "codes":{
      "required": [
        "code",
        "error",
        "message"
      ],
      "properties": {
        "code": {
          "type": "integer"
        },
        "error": {
          "type": "string"
        },
        "message": {
          "type": "integer"
        }
      },
      "example":{
        "properties": "Created",
        "error": "false",
        "message": 100
      }
    }
  }
}